+++ graphql-maven-plugin-logic/src/test/java/com/graphql_java_generator/plugin/generate_code/DocumentParser_Forum_Client_Test.java
@@ -1,6 +1,7 @@
 package com.graphql_java_generator.plugin.generate_code;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -99,30 +100,30 @@ class DocumentParser_Forum_Client_Test {
 		assertEquals("@GraphQLObjectType(\"Topic\")", topic.getAnnotation());
 		int i = 0;
 		checkFieldAnnotation(topic.getFields().get(i++), "id",
-				"@JsonProperty(\"id\")\n\t@GraphQLScalar(fieldName = \"id\", graphQLTypeSimpleName = \"ID\", javaClass = String.class)");
+				"@JsonProperty(\"id\")\n\t@GraphQLScalar(fieldName = \"id\", graphQLTypeSimpleName = \"ID\", javaClass = java.lang.String.class)");
 		checkFieldAnnotation(topic.getFields().get(i++), "date", ""//
 				+ "@JsonProperty(\"date\")\n" //
 				+ "\t@JsonDeserialize(using = CustomJacksonDeserializers.Date.class)\n" //
-				+ "\t@GraphQLScalar(fieldName = \"date\", graphQLTypeSimpleName = \"Date\", javaClass = Date.class)");
+				+ "\t@GraphQLScalar(fieldName = \"date\", graphQLTypeSimpleName = \"Date\", javaClass = java.util.Date.class)");
 		checkFieldAnnotation(topic.getFields().get(i++), "author", "@JsonProperty(\"author\")\n"//
-				+ "\t@GraphQLNonScalar(fieldName = \"author\", graphQLTypeSimpleName = \"Member\", javaClass = Member.class)");
+				+ "\t@GraphQLNonScalar(fieldName = \"author\", graphQLTypeSimpleName = \"Member\", javaClass = org.graphql.mavenplugin.junittest.forum_client_springconfiguration.Member.class)");
 		checkFieldAnnotation(topic.getFields().get(i++), "publiclyAvailable", ""//
 				+ "@JsonProperty(\"publiclyAvailable\")\n"//
-				+ "\t@GraphQLScalar(fieldName = \"publiclyAvailable\", graphQLTypeSimpleName = \"Boolean\", javaClass = Boolean.class)");
+				+ "\t@GraphQLScalar(fieldName = \"publiclyAvailable\", graphQLTypeSimpleName = \"Boolean\", javaClass = java.lang.Boolean.class)");
 		checkFieldAnnotation(topic.getFields().get(i++), "nbPosts", "" //
 				+ "@JsonProperty(\"nbPosts\")\n"//
-				+ "\t@GraphQLScalar(fieldName = \"nbPosts\", graphQLTypeSimpleName = \"Int\", javaClass = Integer.class)");
+				+ "\t@GraphQLScalar(fieldName = \"nbPosts\", graphQLTypeSimpleName = \"Int\", javaClass = java.lang.Integer.class)");
 		checkFieldAnnotation(topic.getFields().get(i++), "title", ""//
 				+ "@JsonProperty(\"title\")\n"//
-				+ "\t@GraphQLScalar(fieldName = \"title\", graphQLTypeSimpleName = \"String\", javaClass = String.class)");
+				+ "\t@GraphQLScalar(fieldName = \"title\", graphQLTypeSimpleName = \"String\", javaClass = java.lang.String.class)");
 		checkFieldAnnotation(topic.getFields().get(i++), "content", ""//
 				+ "@JsonProperty(\"content\")\n"//
-				+ "\t@GraphQLScalar(fieldName = \"content\", graphQLTypeSimpleName = \"String\", javaClass = String.class)");
+				+ "\t@GraphQLScalar(fieldName = \"content\", graphQLTypeSimpleName = \"String\", javaClass = java.lang.String.class)");
 		checkFieldAnnotation(topic.getFields().get(i++), "posts", ""//
 				+ "@JsonProperty(\"posts\")\n"//
 				+ "\t@JsonDeserialize(using = CustomJacksonDeserializers.ListPost.class)\n"//
 				+ "\t@GraphQLInputParameters(names = {\"memberId\", \"memberName\", \"since\"}, types = {\"ID\", \"String\", \"Date\"}, mandatories = {false, false, true}, listDepths = {0, 0, 0}, itemsMandatory = {false, false, false})\n"
-				+ "\t@GraphQLNonScalar(fieldName = \"posts\", graphQLTypeSimpleName = \"Post\", javaClass = Post.class)");
+				+ "\t@GraphQLNonScalar(fieldName = \"posts\", graphQLTypeSimpleName = \"Post\", javaClass = org.graphql.mavenplugin.junittest.forum_client_springconfiguration.Post.class)");
 	}
 
 	/** Tests the annotation. We're in Client mode, thanks to the Spring Configuration used for this test */
@@ -141,7 +142,7 @@ class DocumentParser_Forum_Client_Test {
 		checkFieldAnnotation(mutation.getFields().get(i++), "createBoard", ""//
 				+ "@JsonProperty(\"createBoard\")\n"
 				+ "	@GraphQLInputParameters(names = {\"name\", \"publiclyAvailable\"}, types = {\"String\", \"Boolean\"}, mandatories = {true, false}, listDepths = {0, 0}, itemsMandatory = {false, false})\n"
-				+ "	@GraphQLNonScalar(fieldName = \"createBoard\", graphQLTypeSimpleName = \"Board\", javaClass = Board.class)");
+				+ "	@GraphQLNonScalar(fieldName = \"createBoard\", graphQLTypeSimpleName = \"Board\", javaClass = org.graphql.mavenplugin.junittest.forum_client_springconfiguration.Board.class)");
 	}
 
 	/** Tests the annotation. We're in Client mode, thanks to the Spring Configuration used for this test */
@@ -159,7 +160,7 @@ class DocumentParser_Forum_Client_Test {
 		checkFieldAnnotation(mutation.getFields().get(i++), "createBoard", ""//
 				+ "@JsonProperty(\"createBoard\")\n"
 				+ "\t@GraphQLInputParameters(names = {\"name\", \"publiclyAvailable\"}, types = {\"String\", \"Boolean\"}, mandatories = {true, false}, listDepths = {0, 0}, itemsMandatory = {false, false})\n"
-				+ "\t@GraphQLNonScalar(fieldName = \"createBoard\", graphQLTypeSimpleName = \"Board\", javaClass = Board.class)");
+				+ "\t@GraphQLNonScalar(fieldName = \"createBoard\", graphQLTypeSimpleName = \"Board\", javaClass = org.graphql.mavenplugin.junittest.forum_client_springconfiguration.Board.class)");
 	}
 
 	/** Tests the Data Fetchers that are listed during parsing */
@@ -219,13 +220,15 @@ class DocumentParser_Forum_Client_Test {
 		ObjectType post = (ObjectType) documentParser.getType("Post");
 		assertNotNull(post);
 
-		assertTrue(post.getImports().contains("java.util.Date"), "expecting java.util.Date");
+		// The java class for scalar should not be here. It can lead to name collision, for instance with java.util.Date
+		// and java.sql.Date.
+		// So java class for scalars should used only with full classname (not with imports).
+		assertFalse(post.getImports().contains("java.util.Date"), "expecting java.util.Date");
+
 		assertTrue(post.getImports().contains("com.fasterxml.jackson.annotation.JsonProperty"),
 				"expecting com.fasterxml.jackson.annotation.JsonProperty");
 		assertTrue(post.getImports().contains("com.fasterxml.jackson.databind.annotation.JsonDeserialize"),
 				"expecting com.fasterxml.jackson.databind.annotation.JsonDeserialize");
-		assertTrue(post.getImports().contains("com.graphql_java_generator.GraphQLField"),
-				"expecting com.graphql_java_generator.GraphQLField");
 		assertTrue(post.getImports().contains("com.graphql_java_generator.annotation.GraphQLNonScalar"),
 				"expecting com.graphql_java_generator.annotation.GraphQLNonScalar");
 		assertTrue(post.getImports().contains("com.graphql_java_generator.annotation.GraphQLObjectType"),