+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-server/src/main/java/org/allGraphQLCases/server/impl/DataFetchersDelegateTheSubscriptionTypeImpl.java
@@ -167,4 +167,14 @@ public class DataFetchersDelegateTheSubscriptionTypeImpl implements DataFetchers
 				.map((l) -> Optional.ofNullable(null));
 	}
 
+	/**
+	 * Returns a Flux, that will produce a list of two random dates, every 0.1s
+	 */
+	@Override
+	public Publisher<Optional<List<Date>>> subscribeToAListOfScalar(DataFetchingEnvironment dataFetchingEnvironment) {
+		return Flux//
+				.interval(Duration.ofMillis(100))// A message every 0.1 second
+				.map((l) -> Optional.ofNullable(dataGenerator.generateInstanceList(Date.class, 2)));
+	}
+
 }